+++ CodeWars/Algorithms/AreWeAlternate/Java/src/SolutionTest.java
@@ -1,0 +1,15 @@
+/**
+ * Created by cajetan on 17/07/2017.
+ */
+import org.junit.Test;
+import static org.junit.Assert.assertEquals;
+import org.junit.runners.JUnit4;
+
+public class SolutionTest {
+    @Test
+    public void exampleTests() {
+        assertEquals(true, Solution.isAlt("amazon"));
+        assertEquals(false, Solution.isAlt("apple"));
+        assertEquals(true, Solution.isAlt("banana"));
+    }
+}